requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
pyyaml
python-multipart
loguru
